(93.322 s) -> globalTeardown.*.ts (43.164 s) *AfterAll.test.ts (9.718 s) *Throw.test.ts (18.797 s)
